Adjust README regarding missing config files
Previously, the README mentioned that it is an error if no configuration
files exist. This was true in the very beginning, but after a few
practical tests I came to the conclusion that missing configuration
files should not cause an error, but just be silently ignored. This will
be better for automation scripts that might run on many repositories,
including those without any bump configurations.